+++ taxmate/uae_e_invoicing/report/uae_e_invoice_vat_201_reconciliation/uae_e_invoice_vat_201_reconciliation.js
@@ -1,0 +1,27 @@
+// Copyright (c) 2026, Sourav Singh and contributors
+frappe.query_reports["UAE E-Invoice VAT 201 Reconciliation"] = {
+	filters: [
+		{
+			fieldname: "company",
+			label: __("Company"),
+			fieldtype: "Link",
+			options: "Company",
+			reqd: 1,
+			default: frappe.defaults.get_user_default("Company"),
+		},
+		{
+			fieldname: "from_date",
+			label: __("From Date"),
+			fieldtype: "Date",
+			reqd: 1,
+			default: frappe.datetime.month_start(),
+		},
+		{
+			fieldname: "to_date",
+			label: __("To Date"),
+			fieldtype: "Date",
+			reqd: 1,
+			default: frappe.datetime.month_end(),
+		},
+	],
+};